backend/requirements.txt
fastapi
google-genai
pydantic[all]
python-dotenv
youtube-transcript-api
uvicorn
firebase-admin
python-jose[cryptography]
passlib[bcrypt]
python-multipart